requirements.txt
playwright>=1.30.0
pytest-playwright
google-genai
fastapi
uvicorn>=0.22.0
chromadb>=0.4.17
SpeechRecognition>=3.8.1
pyttsx3>=2.90
